+++ uliweb/utils/common.py
@@ -537,7 +537,37 @@ def get_caller(skip=None):
                     break
             if not found:
                 return filename, lineno, funcname
-            
+
+class classonlymethod(object):
+    """
+    Use to limit the class method can only be called via class object, but not instance
+    object
+    
+    >>> class A(object):
+    ...     @classonlymethod
+    ...     def p(cls):
+    ...         print 'call p()'
+    >>> A.p()
+    call p()
+    >>> a = A()
+    >>> try:
+    ...     a.p()        
+    ... except Exception as e:
+    ...     print e
+    Method p can only be called with class object
+    """
+    def __init__(self, func):
+        self._func = func
+        
+    def __get__(self, model_instance, model_class):
+        from functools import partial, wraps
+        
+        if model_instance:
+            raise Exception("Method %s can only be called with class object" % self._func.__name__)
+        else:
+            f = partial(self._func, model_class)
+            return wraps(self._func)(f)
+
 #if __name__ == '__main__':
 #    log.info('Info: info')
 #    try: